Remove default margin from items inside card
Remove the default margin top of h3 and margin bottom of p
inside card. Remove these to avoid any inconsistencies as the default
of margin of h3 and p may not be the same. Use padding or height
to ensure the space between the content and the edges is always equal.